Erdalcanalkoclar
Erdal Can Alkoclar is a Turkish Inventor, Scientist and Investor. Alkoclar is the co-founder and running president of Mas Pharmaceutials R&D company. He is the sole inventor of 143 applied and 282 pending patents; 44 of them being in current PCT protection. Alkoclar also is the rights holder and applicant of the largest amount of pharmaceutical ingredient patents as sole individual in Turkey. Alkoclar has won a 4th price which is given by Turkish Patent Institute for having the 4th largest number of applied patents in 2014 and The Scientist of the Year Awards for Turkey in 2016 and 2019. Alkoclar is nominated by DMW and EBCA as sole candidate for their Young Scientist of Europe Award of 2020